correcion error en consola - se elimnaron los console.log
Correci on al error que arrojaba la consola por el carProducts.length al intentar revisar el carrito sin haber hecho ninguna selección de productos, elimin

+++ js/cart.js
@@ -1,39 +1,39 @@
 let cartProducts = JSON.parse(localStorage.getItem('cart'));
-console.log(cartProducts)
 const details = document.getElementById('details')
 const printProducts = document.getElementById('cart')
 const totalProducts = document.getElementById('total')
 
 function createCart (){
-    if(cartProducts.length !== 0){
-        for(let i = 0; i< cartProducts.length; i++){
-            const cartSide = cartProducts[i]
-            console.log(cartSide.price);
-            printProducts.innerHTML += `
-            <li id = "list-cart${i}" class = "list-cart">
-                <h2 class = "list-cart-title"><small>${cartSide.name}</small></h2>
-                <img class = "list-cart-image" width = "100" src="${cartSide.image}" alt="">
-                <p class = "list-cart-description"><small>${cartSide.description}</small></p>
-                <p class = "list-cart-price"></p>
-                <p id = "price${i}" class = "list-cart-amount"><small>Price: $ ${cartSide.price}</small></p>
-                <button class="more" onclick = "countMore(${cartSide.id}, ${cartSide.count})"><small>ADD</small></button>
-                <button class="menos" onclick = "countMenos(${cartSide.id}, ${cartSide.count})"><small>REMOVE</small></button>
-            </li>`
-            const total = localStorage.getItem('total')
-            const conteo = localStorage.getItem('conteo')
-            details.innerHTML = `
-            <h3 class = "details-title">DETAILS<h3>
-            <h3 class = "details-total">Total:</h3>
-                <p class = "details-total-number">$ ${total}</p> 
-            <h3 class = "details-amount">Products Amount:</h3>
-                <p class = "details-amount-number">${conteo}</p>
-                <button class="delete-car" onclick = "deleteCar()">DELETE CAR</button>
-            `
-            const acumulado = document.getElementsByTagName('li')[i].querySelector('.list-cart-price')
-            console.log(acumulado);
-            acumulado.innerHTML = `Amount: ${cartProducts[i].count}`
+    if(cartProducts){
+        if(cartProducts.length !== 0){
+            for(let i = 0; i< cartProducts.length; i++){
+                const cartSide = cartProducts[i]
+                printProducts.innerHTML += `
+                <li id = "list-cart${i}" class = "list-cart">
+                    <h2 class = "list-cart-title"><small>${cartSide.name}</small></h2>
+                    <img class = "list-cart-image" width = "100" src="${cartSide.image}" alt="">
+                    <p class = "list-cart-description"><small>${cartSide.description}</small></p>
+                    <p class = "list-cart-price"></p>
+                    <p id = "price${i}" class = "list-cart-amount"><small>Price: $ ${cartSide.price}</small></p>
+                    <button class="more" onclick = "countMore(${cartSide.id}, ${cartSide.count})"><small>ADD</small></button>
+                    <button class="menos" onclick = "countMenos(${cartSide.id}, ${cartSide.count})"><small>REMOVE</small></button>
+                </li>`
+                const total = localStorage.getItem('total')
+                const conteo = localStorage.getItem('conteo')
+                details.innerHTML = `
+                <h3 class = "details-title">DETAILS<h3>
+                <h3 class = "details-total">Total:</h3>
+                    <p class = "details-total-number">$ ${total}</p> 
+                <h3 class = "details-amount">Products Amount:</h3>
+                    <p class = "details-amount-number">${conteo}</p>
+                    <button class="delete-car" onclick = "deleteCar()">DELETE CAR</button>
+                `
+                const acumulado = document.getElementsByTagName('li')[i].querySelector('.list-cart-price')
+                acumulado.innerHTML = `Amount: ${cartProducts[i].count}`
             }
+        }
     }
+    
 }
 function countMore (id, count){
     for(let i = 0; i< cartProducts.length; i++){
@@ -46,7 +46,6 @@ function countMore (id, count){
             localStorage.setItem('conteo', conteo)
             localStorage.setItem('total', total)
             localStorage.setItem('cart', JSON.stringify(cartProducts))
-            console.log(`el contador de ${cartProducts[i].name}  ${cartProducts[i].count}`)
             details.innerHTML = `
             <h3 class = "details-title">DETAILS<h3>
             <h3 class = "details-total">Total:</h3>
@@ -56,9 +55,7 @@ function countMore (id, count){
                 <button class="delete-car" onclick = "deleteCar()">DELETE CAR</button>
             `
             const acumulado = document.getElementsByTagName('li')[i].querySelector('.list-cart-price')
-            console.log(acumulado);
             acumulado.innerHTML = `Amount: ${cartProducts[i].count}`
-            
         }
     }
 }
@@ -83,20 +80,16 @@ function countMenos (id, count){
                 <button class="delete-car" onclick = "deleteCar()">DELETE CAR</button>
             `
             const acumulado = document.getElementsByTagName('li')[i].querySelector('.list-cart-price')
-            console.log(acumulado);
             acumulado.innerHTML = `Amount: ${cartProducts[i].count}`
 
-            console.log(`el contador de ${cartProducts[i].name}  ${cartProducts[i].count}`)
             if(cartProducts[i].count<1){
                 cartProducts.splice(i, 1)
-                console.log(cartProducts);
                 localStorage.setItem('cart', JSON.stringify(cartProducts))
                 const padre = document.getElementById('cart')
                 const hijo = document.getElementsByTagName("li")[i]
                 const removido = padre.removeChild(hijo)
                 if(cartProducts.length===0){
                     localStorage.removeItem('cart')
-                    console.log('el arreglo esta vacio')
                     printProducts.innerHTML = `
                     <i class="material-icons md-48" id="sad-icon">sentiment_very_dissatisfied</i>
                     <h3>Your cart is empty click <button><a class="back-button" href="index.html">HERE!</a></button></h3>
@@ -108,11 +101,9 @@ function countMenos (id, count){
 }
 
 function deleteCar(){
-    console.log('este es el boton de eliminar ');
     const cartProducts = ''
     const total = 0
     const conteo = 0
-    console.log(printProducts);
     localStorage.setItem('cart', cartProducts)
     localStorage.setItem('conteo', conteo)
     localStorage.setItem('total', total)
